# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the documentation for the aanwezigheidstool API.",
        "version": "v1.0.0",
        "title": "Aanwezigheidstool",
        "license": {
            "name": "Our github",
            "url": "https://github.com/Rac-Software-Development/werkplaats-3-rest-alien-software"
        }
    },
    "host": "127.0.0.1:81",
    "basePath": "/api",
...
    "schemes": [
        "http"
    ],
    "paths": {
        "/meeting": {
            "get": {
                "tags": [
                    "meeting"
                ],
                "summary": "get all meetings",
                "description": "",
                "operationId": "getMeetings",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            },
            "post": {
                "tags": [
                    "meeting"
                ],
                "summary": "create a new meeting",
                "description": "",
                "operationId": "createMeetings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "date",
                        "in": "path",
                        "description": "A date",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "path",
                        "description": "A name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "path",
                        "description": "A description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "start_time",
                        "in": "path",
                        "description": "A start time",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "end_time",
                        "in": "path",
                        "description": "A end time",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
        "/meeting/{id}": {
            "get": {
                "tags": [
                    "meeting"
                ],
                "summary": "get a specific meeting",
                "description": "",
                "operationId": "findMeetingById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of meeting to retrieve",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            },
            "delete": {
                "tags": [
                    "meeting"
                ],
                "summary": "delete a specific meeting",
                "description": "",
                "operationId": "deleteMeetingById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of meeting to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            },
            "put": {
                "tags": [
                    "meeting"
                ],
                "summary": "put a specific meeting",
                "description": "",
                "operationId": "UpdateMeetingById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of meeting to put",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "date",
                        "in": "body",
                        "description": "a date",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "body",
                        "description": "a name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "body",
                        "description": "a description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "start_time",
                        "in": "body",
                        "description": "a start time",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "end_time",
                        "in": "body",
                        "description": "an end time",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            },
            "patch": {
                "tags": [
                    "meeting"
                ],
                "summary": "patch a specific meeting",
                "description": "",
                "operationId": "UpdateMeetingById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of meeting to patch",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "date",
                        "in": "body",
                        "description": "a date",
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "body",
                        "description": "a name",
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "body",
                        "description": "a description",
                        "type": "string"
                    },
                    {
                        "name": "start_time",
                        "in": "body",
                        "description": "a start time",
                        "type": "string"
                    },
                    {
                        "name": "end_time",
                        "in": "body",
                        "description": "an end time",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
        "/meeting/between/{start}/{end}": {
            "get": {
                "tags": [
                    "meeting"
                ],
                "summary": "get a specific meetings between a start and end date",
                "description": "",
                "operationId": "findMeetingById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "start",
                        "in": "path",
                        "description": "begin date of meetings between two dates",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "end",
                        "in": "path",
                        "description": "end date of meetings between two dates",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
        "/meeting/filter/{filter}": {
            "get": {
                "tags": [
                    "meeting"
                ],
                "summary": "get list of meetings by name",
                "description": "",
                "operationId": "findMeetingByName",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filter",
                        "in": "path",
                        "description": "a part of a name of a meeting",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
        "/meeting/limit/{limit}": {
            "get": {
                "tags": [
                    "meeting"
                ],
                "summary": "get a list of meetings from today or later limited by a number",
                "description": "",
                "operationId": "findMeetingByName",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "a limit for the amount of returned meetings",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
        "/studentmeeting/{code}": {
            "get": {
                "tags": [
                    "studentmeeting"
                ],
                "summary": "returns list of students that are present to the meeting of this meeting code",
                "description": "",
                "operationId": "studentsOfMeeting",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "meeting_code of a meeting",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
        "/groupmeeting": {
            "post": {
                "tags": [
                    "groupmeeting"
                ],
                "summary": "add a group to a meeting",
                "description": "",
                "operationId": "groupToMeeting",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "body",
                        "description": "id of a group",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "meeting_id",
                        "in": "body",
                        "description": "id of a meeting",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
...
        "/student/{id}": {
            "get": {
                "tags": [
                    "student"
                ],
                "summary": "get a student by id",
                "description": "",
                "operationId": "getStudentById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of student",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "student"
                ],
                "summary": "delete student by id",
                "description": "",
                "operationId": "deleteStudent",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of student to delete",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/ApiResponse"
                        }
                    }
                }
 
            }
        },
...
    },
    "definitions": {
        "ApiResponse": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "array",
                    "example": []
                },
                "error": {
                    "type": "string",
                    "example": ""
                }
            }
        },
...
    }
}